refactor api_server, split endpoints to routers
