+++ neurons/validator/api_server/rate_limiter.py
@@ -1,0 +1,5 @@
+# neurons/validator/api_server/utils/rate_limiter.py
+from slowapi import Limiter
+from slowapi.util import get_remote_address
+
+limiter = Limiter(key_func=get_remote_address)
